Add auth for remaining apps
- hide plan until more stable

+++ convex/schedules.ts
@@ -1,20 +1,67 @@
-import { mutation, query } from "./_generated/server";
 import { v } from "convex/values";
+import { mutation, query } from "./_generated/server";
 
 export const list = query(async (ctx) => {
-  return await ctx.db.query("schedules").collect();
+  const identity = await ctx.auth.getUserIdentity();
+  if (!identity) {
+    throw new Error("Unauthenticated call to mutation");
+  }
+  const user = await ctx.db
+    .query("users")
+    .withIndex("by_token", (q) =>
+      q.eq("tokenIdentifier", identity.tokenIdentifier),
+    )
+    .unique();
+  if (!user) {
+    throw new Error("Unauthenticated call to mutation");
+  }
+  return await ctx.db
+    .query("schedules")
+    .filter((q) => q.eq(q.field("userId"), user._id))
+    .collect();
 });
 
 export const listTemplates = query(async (ctx) => {
+  const identity = await ctx.auth.getUserIdentity();
+  if (!identity) {
+    throw new Error("Unauthenticated call to mutation");
+  }
+  const user = await ctx.db
+    .query("users")
+    .withIndex("by_token", (q) =>
+      q.eq("tokenIdentifier", identity.tokenIdentifier),
+    )
+    .unique();
+  if (!user) {
+    throw new Error("Unauthenticated call to mutation");
+  }
   return await ctx.db
     .query("schedules")
-    .filter((q) => q.eq(q.field("isTemplate"), true))
+    .filter((q) =>
+      q.and(
+        q.eq(q.field("isTemplate"), true),
+        q.eq(q.field("userId"), user._id),
+      ),
+    )
     .collect();
 });
 
 export const get = query({
   args: { scheduleId: v.id("schedules") },
   async handler(ctx, { scheduleId }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     const schedule = await ctx.db.get(scheduleId);
 
     if (!schedule) {
@@ -28,9 +75,27 @@ export const get = query({
 export const getOrderedActivities = query({
   args: { scheduleId: v.id("schedules") },
   async handler(ctx, { scheduleId }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     const activities = await ctx.db
       .query("activities")
-      .filter((q) => q.eq(q.field("scheduleId"), scheduleId))
+      .filter((q) =>
+        q.and(
+          q.eq(q.field("scheduleId"), scheduleId),
+          q.eq(q.field("userId"), user._id),
+        ),
+      )
       .collect();
 
     if (activities.length === 0) {
@@ -46,9 +111,24 @@ export const getByDate = query({
     date: v.string(),
   },
   async handler(ctx, { date }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     return await ctx.db
       .query("schedules")
-      .withIndex("by_date", (q) => q.eq("date", date))
+      .filter((q) =>
+        q.and(q.eq("date", date), q.eq(q.field("userId"), user._id)),
+      )
       .first();
   },
 });
@@ -61,11 +141,25 @@ export const createSchedule = mutation({
     length: v.number(),
   },
   async handler(ctx, { name, date, isTemplate, length }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     return await ctx.db.insert("schedules", {
       name: name,
       date: date,
       isTemplate: isTemplate,
       length: length,
+      userId: user._id,
     });
   },
 });
@@ -80,11 +174,26 @@ export const updateSchedule = mutation({
     start: v.optional(v.number()),
   },
   async handler(ctx, { scheduleId, name, date, isTemplate, length, start }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     return await ctx.db.patch(scheduleId, {
       name: name,
       date: date,
       isTemplate: isTemplate,
       length: length,
+      updated: Date.now(),
+      userId: user._id,
     });
   },
 });
@@ -94,15 +203,46 @@ export const deleteSchedule = mutation({
     scheduleId: v.id("schedules"),
   },
   async handler(ctx, { scheduleId }) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     return await ctx.db.delete(scheduleId);
   },
 });
 
 export const templateSchedules = query({
   async handler(ctx) {
+    const identity = await ctx.auth.getUserIdentity();
+    if (!identity) {
+      throw new Error("Unauthenticated call to mutation");
+    }
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) =>
+        q.eq("tokenIdentifier", identity.tokenIdentifier),
+      )
+      .unique();
+    if (!user) {
+      throw new Error("Unauthenticated call to mutation");
+    }
     return await ctx.db
       .query("schedules")
-      .filter((q) => q.eq(q.field("isTemplate"), true))
+      .filter((q) =>
+        q.and(
+          q.eq(q.field("isTemplate"), true),
+          q.eq(q.field("userId"), user._id),
+        ),
+      )
       .collect();
   },
 });
